♻️ move createHandlingStack to the tools folder
