+++ packages/core/src/domain/error/trackRuntimeError.ts
@@ -1,8 +1,8 @@
 import { instrumentMethod } from '../../tools/instrumentMethod'
 import type { Observable } from '../../tools/observable'
 import { clocksNow } from '../../tools/utils/timeUtils'
-import type { StackTrace } from './computeStackTrace'
-import { computeStackTrace, computeStackTraceFromOnErrorMessage } from './computeStackTrace'
+import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
+import { computeStackTrace, computeStackTraceFromOnErrorMessage } from '../../tools/stackTrace/computeStackTrace'
 import { computeRawError } from './error'
 import type { RawError } from './error.types'
 import { ErrorHandling, ErrorSource, NonErrorPrefix } from './error.types'
