+++ packages/core/src/domain/telemetry/telemetry.ts
@@ -1,6 +1,7 @@
 import type { Context } from '../../tools/serialisation/context'
 import { ConsoleApiName } from '../../tools/display'
-import { toStackTraceString, NO_ERROR_STACK_PRESENT_MESSAGE } from '../error/error'
+import { NO_ERROR_STACK_PRESENT_MESSAGE } from '../error/error'
+import { toStackTraceString } from '../../tools/stackTrace/handlingStack'
 import { getExperimentalFeatures } from '../../tools/experimentalFeatures'
 import type { Configuration } from '../configuration'
 import { INTAKE_SITE_STAGING, INTAKE_SITE_US1_FED } from '../configuration'
@@ -13,8 +14,8 @@ import { performDraw } from '../../tools/utils/numberUtils'
 import { jsonStringify } from '../../tools/serialisation/jsonStringify'
 import { combine } from '../../tools/mergeInto'
 import { NonErrorPrefix } from '../error/error.types'
-import type { StackTrace } from '../error/computeStackTrace'
-import { computeStackTrace } from '../error/computeStackTrace'
+import type { StackTrace } from '../../tools/stackTrace/computeStackTrace'
+import { computeStackTrace } from '../../tools/stackTrace/computeStackTrace'
 import { getConnectivity } from '../connectivity'
 import type { TelemetryEvent } from './telemetryEvent.types'
 import type {
