+++ packages/core/src/tools/stackTrace/capturedExceptions.specHelper.ts
