+++ packages/core/src/tools/stackTrace/computeStackTrace.ts
@@ -4,7 +4,7 @@
  * Reference implementation: https://github.com/csnover/TraceKit/blob/04530298073c3823de72deb0b97e7b38ca7bcb59/tracekit.js
  */
 
-import { startsWith } from '../../tools/utils/polyfills'
+import { startsWith } from '../utils/polyfills'
 
 export interface StackFrame {
   url?: string
